Notify clients when a message is deleted.
When connecting to the `/stream` endpoint to the server, clients can
specify an API version. Per default, this version is 1, which
essentially filters all events but the Message event, and provides the
clients with a backwards-compatible JSON message.

We pass internal messages to the client handler, which will then pass
the messages to an event encoder. The encoder returns the right format
depending on the API version of the client.

When deleting messages, we retrieve the messages that are to be deleted,
and box them inside a MessageDeletions struct. This struct implements
the event interface, so that the event encoder can return an appropriate
format to the client handler.

The client handler will then notify all the clients with an API version
of at least 2 of the new event.