Fixed error net:ERR_INCOMPLETE_CHUNKED_ENCODING
If compression is enabled, go does not add a Content-Length Header to the request.
This causes problems with transfers in some browsers.
It is nicely visible while doing e2e tests with:
cd ui
yarn build
yarn run test user

This is maybe a bug in Chrome:
https://issues.chromium.org/issues/40406950

If you want compression use a proxy before gotify that handles the compression and headers correctly.

For reference:
https://github.com/golang/go/issues/19420
https://go-review.googlesource.com/c/go/+/381956